ai-engineer/05-multimodal-ai/speech-to-summary/requirements.txt
openai-whisper
langchain-core
langchain-ollama
